OpenID4VP: Move type aliases, helper functions to correct package
