+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/RemoteResourceRetrieverFunction.kt
@@ -1,0 +1,7 @@
+package at.asitplus.wallet.lib
+
+/**
+ * Implementations need to fetch the url passed in, and return either the body, if there is one,
+ * or the HTTP header `Location`, i.e. if the server sends the request object as a redirect.
+ */
+typealias RemoteResourceRetrieverFunction = suspend (String) -> String?
