+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/RequestObjectJwsVerifier.kt
@@ -1,0 +1,11 @@
+package at.asitplus.wallet.lib.oidc
+
+import at.asitplus.openid.RequestParameters
+import at.asitplus.signum.indispensable.josef.JwsSigned
+
+/**
+ * Implementations need to verify the passed [at.asitplus.signum.indispensable.josef.JwsSigned] and return its result
+ */
+fun interface RequestObjectJwsVerifier {
+    operator fun invoke(jws: JwsSigned<RequestParameters>): Boolean
+}
